requirements.txt
Flask
flask-cors
pymongo
python-dotenv
opencv-python
Pillow
torch
torchvision
gdown
pyserial